+++ src/main/java/com/klarna/hiverunner/HiveServerContainer.java
@@ -19,10 +19,10 @@ import com.google.common.base.Function;
 import com.google.common.base.Joiner;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Iterables;
-import com.klarna.hiverunner.sql.StatementsSplitter;
 import org.apache.hadoop.hive.conf.HiveConf;
-import org.apache.hadoop.hive.ql.exec.tez.TezJobMonitor;
-import org.apache.hadoop.hive.ql.parse.VariableSubstitution;
+import org.apache.hadoop.hive.conf.HiveVariableSource;
+import org.apache.hadoop.hive.conf.VariableSubstitution;
+import org.apache.hadoop.hive.ql.exec.tez.TezJobExecHelper;
 import org.apache.hadoop.hive.ql.session.SessionState;
 import org.apache.hive.service.Service;
 import org.apache.hive.service.cli.CLIService;
@@ -146,17 +146,6 @@ public class HiveServerContainer {
     }
 
     /**
-     * Executes a hive script.
-     *
-     * @param hiveql hive script statements.
-     */
-    public void executeScript(String hiveql) {
-        for (String statement : StatementsSplitter.splitStatements(hiveql)) {
-            executeStatement(statement);
-        }
-    }
-
-    /**
      * Release all resources.
      * <p>
      * This call will never throw an exception as it makes no sense doing that in the tear down phase.
@@ -166,7 +155,7 @@ public class HiveServerContainer {
 
 
         try {
-            TezJobMonitor.killRunningJobs();
+          TezJobExecHelper.killRunningJobs();
         } catch (Throwable e) {
             LOGGER.warn("Failed to kill tez session: " + e.getMessage() + ". Turn on log level debug for stacktrace");
             LOGGER.debug(e.getMessage(), e);
@@ -174,7 +163,7 @@ public class HiveServerContainer {
 
         try {
             // Reset to default schema
-            executeScript("USE default;");
+            executeStatement("USE default");
         } catch (Throwable e) {
             LOGGER.warn("Failed to reset to default schema: " + e.getMessage() +
                     ". Turn on log level debug for stacktrace");
@@ -219,7 +208,15 @@ public class HiveServerContainer {
         // Make sure to set the session state for this thread before returning the VariableSubstitution. If not set,
         // hivevar:s will not be evaluated.
         SessionState.setCurrentSessionState(currentSessionState);
-        return new VariableSubstitution();
+
+        final SessionState ss = currentSessionState;
+        return new VariableSubstitution(new HiveVariableSource() {
+            @Override
+            public Map<String, String> getHiveVariable() {
+                return ss.getHiveVariables();
+            }
+        });
+
     }
 }