+++ src/main/java/com/klarna/hiverunner/TimeoutException.java
@@ -17,6 +17,8 @@ package com.klarna.hiverunner;
 
 public class TimeoutException extends RuntimeException {
 
+    private static final long serialVersionUID = 1L;
+
     public TimeoutException() {
         super();
     }