+++ src/test/java/com/klarna/hiverunner/ReservedKeywordTest.java
@@ -16,15 +16,11 @@
 package com.klarna.hiverunner;
 
 import com.klarna.hiverunner.annotations.HiveSQL;
-import org.junit.Assert;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import java.io.IOException;
-import java.io.OutputStream;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.UUID;
 
 @RunWith(StandaloneHiveRunner.class)
 public class ReservedKeywordTest {
@@ -39,6 +35,7 @@ public class ReservedKeywordTest {
      * 'hive.support.sql11.reserved.keywords' to false.
      */
     @Test
+    @Ignore("Since Hive 2.3.0 this property is no longer available in hive see https://issues.apache.org/jira/browse/HIVE-14872, use backticks")
     public void reservedKeywordsShouldBeAllowedWhenHiveConfIsSet() throws IOException {
 
         hiveShell.setHiveConfValue("hive.support.sql11.reserved.keywords", "false");